Define a logging library for the Astarte Device SDK.
Before, all the logging was made using the System.out.println or the System.out.error functions causing an unwanted and difficult to regulate info-dump in production logs.
The adoption of standard logging methods was a necessary operation, but we had to face another problem: the absence of a single standard between Android and Oracle Java.
We resolved it by implementing the logging class twice, one for each standard in each platform-specific module, and by creating a proxy class in the common module to handle the right implementation inside said common module. The proxy class handles the burden of instantiate the correct implementation using the reflection property of Java.

On top of that, in the Java implementation, a VM Property is exposed to let the user to set the preferred logging level.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>